# src/management-service/docs/swagger.yaml
definitions:
...
  request.CreateForum:
    properties:
      categoryId:
        type: string
      description:
        type: string
      forumName:
        type: string
      type:
        type: string
    type: object
...
  request.ForumUpdateCategoryOwner:
    properties:
      categoryId:
        type: string
      forumId:
        type: string
    type: object
  request.ForumUpdateDescription:
    properties:
      description:
        type: string
      forumId:
        type: string
    type: object
  request.ForumUpdateName:
    properties:
      forumId:
        type: string
      name:
        type: string
    type: object
  response.Standard:
    properties:
      data: {}
      message:
        type: string
      statusCode:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /create-forum:
    post:
      consumes:
      - application/json
      description: Create a new forum with a name, description, and category ID
      parameters:
      - description: Forum creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateForum'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Category Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "409":
          description: Conflict (Duplicate Forum)
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Create a new forum
      tags:
      - forums
...
  /delete-forum/{id}:
    delete:
      description: Soft delete a forum by its UUID
      parameters:
      - description: Forum ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Forum Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Delete a forum
      tags:
      - forums
...
  /get-all-forums:
    get:
      description: Get a list of all forums
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Get all forums
      tags:
      - forums
...
  /get-forum-id/{id}:
    get:
      description: Get a single forum by its UUID
      parameters:
      - description: Forum ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Get forum by ID
      tags:
      - forums
...
  /restore-forum/{id}:
    post:
      description: Restore a forum that was previously soft deleted by its UUID
      parameters:
      - description: Forum ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Forum Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Restore a soft deleted forum
      tags:
      - forums
...
  /update-forum-category:
    put:
      consumes:
      - application/json
      description: Update the category of a forum by its UUID and new category ID
      parameters:
      - description: Forum and Category IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ForumUpdateCategoryOwner'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Forum or Category Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Update the category owner of a forum
      tags:
      - forums
  /update-forum-description:
    put:
      consumes:
      - application/json
      description: Update the description of a forum by its UUID
      parameters:
      - description: Forum description update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ForumUpdateDescription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Forum Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Update the description of a forum
      tags:
      - forums
  /update-forum-name:
    put:
      consumes:
      - application/json
      description: Update the name of a forum by its UUID
      parameters:
      - description: Forum name update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ForumUpdateName'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Standard'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/response.Standard'
        "404":
          description: Forum Not Found
          schema:
            $ref: '#/definitions/response.Standard'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Standard'
      summary: Update the name of a forum
      tags:
      - forums
...
swagger: "2.0"